Extract EqualityTypeSpecifyingHelper from TypeSpecifier
resolveEqual and resolveIdentical (plus their private helpers
resolveNormalizedIdentical, findTypeExpressionsFromBinaryOperation,
specifyTypesForConstantBinaryExpression and
specifyTypesForConstantStringBinaryExpression) were only used by
BinaryOpHandler. Move them into a new AutowiredService
EqualityTypeSpecifyingHelper and rename the entry points to
specifyTypesForEqual()/specifyTypesForIdentical() to signal they only
produce SpecifiedTypes.
