+++ phpstan-baseline.neon
@@ -43,6 +43,18 @@ parameters:
 			path: src/Analyser/ExprHandler/BooleanNotHandler.php
 
 		-
+			rawMessage: 'Doing instanceof PHPStan\Type\ConstantScalarType is error-prone and deprecated. Use Type::isConstantScalarValue() or Type::getConstantScalarTypes() or Type::getConstantScalarValues() instead.'
+			identifier: phpstanApi.instanceofType
+			count: 2
+			path: src/Analyser/ExprHandler/Helper/EqualityTypeSpecifyingHelper.php
+
+		-
+			rawMessage: 'Doing instanceof PHPStan\Type\Constant\ConstantBooleanType is error-prone and deprecated. Use Type::isTrue() or Type::isFalse() instead.'
+			identifier: phpstanApi.instanceofType
+			count: 3
+			path: src/Analyser/ExprHandler/Helper/EqualityTypeSpecifyingHelper.php
+
+		-
 			rawMessage: 'Doing instanceof PHPStan\Type\Constant\ConstantStringType is error-prone and deprecated. Use Type::getConstantStrings() instead.'
 			identifier: phpstanApi.instanceofType
 			count: 2
@@ -103,15 +115,9 @@ parameters:
 			path: src/Analyser/RuleErrorTransformer.php
 
 		-
-			rawMessage: 'Doing instanceof PHPStan\Type\ConstantScalarType is error-prone and deprecated. Use Type::isConstantScalarValue() or Type::getConstantScalarTypes() or Type::getConstantScalarValues() instead.'
-			identifier: phpstanApi.instanceofType
-			count: 2
-			path: src/Analyser/TypeSpecifier.php
-
-		-
 			rawMessage: 'Doing instanceof PHPStan\Type\Constant\ConstantBooleanType is error-prone and deprecated. Use Type::isTrue() or Type::isFalse() instead.'
 			identifier: phpstanApi.instanceofType
-			count: 5
+			count: 2
 			path: src/Analyser/TypeSpecifier.php
 
 		-
